+++ src/controllers/recognition.ts
@@ -1,6 +1,7 @@
 import { AppDataSource } from '@/data-source';
 import { Recognition } from '@/entities/';
 import logger from '@/utils/logger';
+import { In } from 'typeorm';
 import { InstallationController, UserController, WalletController } from './';
 
 export type RecognitionSummary = {
@@ -20,6 +21,7 @@ export class RecognitionController {
   private readonly recognitionRepository =
     AppDataSource.getRepository(Recognition);
   private readonly userController = new UserController();
+  private readonly walletController = new WalletController();
 
   public async save(params: SaveRecognitionParams) {
     const { toId, teamId, fromId, botToken } = params;
@@ -41,7 +43,7 @@ export class RecognitionController {
         });
       }
 
-      const response = await this.recognitionRepository.save({
+      const { id } = await this.recognitionRepository.save({
         fromId,
         fromName: fromUser.name,
         toId,
@@ -51,14 +53,14 @@ export class RecognitionController {
       });
 
       const { defaultAmount } = await new InstallationController().find(teamId);
-      if (response.id) {
-        await new WalletController().deposit({
+      if (id) {
+        await this.walletController.deposit({
           ownerId: toId,
           amount: defaultAmount,
           teamId,
         });
       }
-      return { ok: true };
+      return { ok: true, id };
     } catch (error) {
       logger.error('RecognitionController.save()', { error });
       return { ok: false };
@@ -91,5 +93,68 @@ export class RecognitionController {
       .limit(20)
       .getRawMany();
     return summary;
+  }
+
+  public async update({
+    teamId,
+    id,
+    params,
+  }: {
+    teamId: string;
+    id: number[];
+    params: Partial<Recognition>;
+  }): Promise<void> {
+    await this.recognitionRepository.update({ teamId, id: In(id) }, params);
+  }
+
+  public async find({
+    teamId,
+    params,
+  }: {
+    teamId: string;
+    params: Partial<Recognition>;
+  }): Promise<Recognition[]> {
+    return this.recognitionRepository.find({
+      where: {
+        teamId,
+        ...params,
+      },
+    });
+  }
+
+  public async delete({
+    teamId,
+    params,
+  }: {
+    teamId: string;
+    params: Partial<Recognition>;
+  }) {
+    try {
+      const wallets = await this.recognitionRepository.find({
+        where: {
+          teamId,
+          ...params,
+        },
+      });
+
+      const { defaultAmount } = await new InstallationController().find(teamId);
+
+      await Promise.all(
+        wallets.map((wallet) =>
+          this.walletController.withdraw({
+            teamId,
+            ownerId: wallet.toId,
+            amount: defaultAmount,
+          }),
+        ),
+      );
+
+      await this.recognitionRepository.delete({
+        teamId,
+        ...params,
+      });
+    } catch (error) {
+      logger.error('RecognitionController.delete()', { error });
+    }
   }
 }