+++ src/listeners/views/index.ts
@@ -1,4 +1,5 @@
 import { App } from '@slack/bolt';
+import cancelKudosCallback from './cancel-kudos';
 import generateGiftCardCallback from './generate-card';
 import generatePrizesReportCallback from './generate-prizes-report';
 import giveKudosViewCallback from './give-kudos';
@@ -9,6 +10,7 @@ const register = (app: App) => {
   app.view('generate_gift_card', generateGiftCardCallback);
   app.view('settings_view', saveSettingsCallback);
   app.view('generate_prizes_report', generatePrizesReportCallback);
+  app.view('cancel_selected_kudos', cancelKudosCallback);
 };
 
 export default { register };